Add command line utility to call post-harvest tasks

+++ dlme_airflow/utils/qnl.py
@@ -11,9 +11,12 @@ def merge_records(**kwargs):
     root_dir = os.path.dirname(os.path.abspath("metadata"))
     data_path = coll.data_path()
     working_csv = os.path.join(root_dir, "metadata", data_path, "data.csv")
-    df = read_csv_with_lists(working_csv)
-    df = merge_df(df)
-    df.to_csv(working_csv)
+    if os.path.isfile(working_csv):
+        df = read_csv_with_lists(working_csv)
+        df = merge_df(df)
+        df.to_csv(working_csv)
+
+    return working_csv
 
 
 def read_csv_with_lists(path) -> pandas.DataFrame:
