+++ dlme_airflow/utils/yale.py
@@ -25,8 +25,11 @@ def remove_babylonian_non_relevant(**kwargs):
     root_dir = os.path.dirname(os.path.abspath("metadata"))
     data_path = coll.data_path()
     working_csv = os.path.join(root_dir, "working", data_path, "data.csv")
-    df = pd.read_csv(working_csv)
-    # Filter out non relevant records and over write the csv
-    df = df[~df["geographic_country"].isin(NON_RELEVANT_COUNTRIES)]
 
-    df.to_csv(working_csv)
+    if os.path.isfile(working_csv):
+        df = pd.read_csv(working_csv)
+        # Filter out non relevant records and over write the csv
+        df = df[~df["geographic_country"].isin(NON_RELEVANT_COUNTRIES)]
+        df.to_csv(working_csv)
+
+    return working_csv
